Use git submodules to manage annotationframework dependency
Since annotationframework plugins are not published to grails repo, we
need a way to manage the dependency and lock the dependency
version/commit as well. Using git submodules for now, but they can be
moved to more standard grails way in the future once annotationframework
is published to grails repo.